refactor: move free-standing cli functions to structs
