+++ cloud-checksum/src/io/copy/mod.rs
@@ -2,12 +2,12 @@
 //!
 
 use crate::checksum::Ctx;
+use crate::cli::MetadataCopy;
 use crate::error::Error::CopyError;
 use crate::error::Result;
 use crate::io::copy::aws::S3Builder;
 use crate::io::copy::file::FileBuilder;
 use crate::io::{default_s3_client, Provider};
-use crate::MetadataCopy;
 use aws_sdk_s3::Client;
 use dyn_clone::DynClone;
 use std::collections::HashMap;
