+++ cloud-checksum/src/main.rs
@@ -1,12 +1,11 @@
-use cloud_checksum::cli::execute_args;
+use cloud_checksum::cli::Command;
 use cloud_checksum::error::Result;
-use cloud_checksum::Command;
 
 #[tokio::main]
 async fn main() -> Result<()> {
     let args = Command::parse_args()?;
 
-    execute_args(args).await?;
+    args.execute().await?;
 
     Ok(())
 }
